pyproject.toml
[project]
name = "log-essence"
version = "0.1.0b1"
description = "Intelligent log analysis for LLMs - extract patterns, redact secrets, compress 80-95%"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Pete" }
]
keywords = [
    "mcp",
    "llm",
    "logs",
    "log-analysis",
    "log-parser",
    "claude",
    "ai",
    "devtools",
    "observability",
    "template-extraction",
    "drain3",
    "semantic-clustering",
    "secret-detection",
    "pii-redaction",
    "docker-logs",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Typing :: Typed",
]
dependencies = [
    "fastmcp>=0.4.0",
    "drain3>=0.9.11",
    "fastembed>=0.4.0",
    "numpy>=1.26.0",
    "tiktoken>=0.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
demo = [
    "playwright>=1.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
ui = [
    "streamlit>=1.40.0",
]

[project.urls]
Homepage = "https://github.com/petebytes/log-essence"
Repository = "https://github.com/petebytes/log-essence"
Issues = "https://github.com/petebytes/log-essence/issues"

[project.scripts]
log-essence = "log_essence.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/log_essence"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
